requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.11
discord-webhook==0.14.0
idna==3.3
oauthlib==3.2.0
python-dotenv==0.19.2
requests==2.27.1
requests-oauthlib==1.3.1
tweepy==4.5.0
urllib3==1.26.8
